replace old plan with new plan
