+++ src/views/Plan/ClearDay.tsx
