+++ src/views/Plan/Day.tsx
@@ -1,26 +1,57 @@
 import { type Dinner } from "@prisma/client";
-import { cn } from "../../lib/utils";
-import { DialogTrigger } from "@radix-ui/react-dialog";
 import { format } from "date-fns";
+import { cn } from "../../lib/utils";
+import { Dialog, DialogTrigger } from "../../components/ui/dialog";
+import { PlannedDinner } from "./PlannedDinner";
+import { type DinnerWithTags } from "../../utils/types";
+import { useState } from "react";
+import { PlanDay } from "./PlanDay";
 
-type DayProps = {
+type Props = {
   date: Date;
-  setSelectedDay: (date: Date) => void;
-  plannedDinner?: Dinner;
+  plannedDinner?: DinnerWithTags;
 };
 
-export const Day = ({ date, setSelectedDay, plannedDinner }: DayProps) => {
+export const Day = ({ date, plannedDinner }: Props) => {
+  const [dialogOpen, setDialogOpen] = useState(false);
+  const [changePlan, setChangePlan] = useState(!plannedDinner);
+
+  const onOpenChange = (open: boolean) => {
+    setDialogOpen(open);
+    if (!open) {
+      setChangePlan(false);
+    }
+  };
+
   return (
-    <DialogTrigger asChild onClick={() => setSelectedDay(date)}>
-      <div
-        className={cn(
-          "flex cursor-pointer flex-col rounded border px-2 py-2 hover:bg-slate-100",
+    <Dialog open={dialogOpen} onOpenChange={onOpenChange}>
+      <DialogTrigger asChild>
+        <div
+          className={cn(
+            "flex cursor-pointer flex-col rounded border p-2 hover:bg-slate-100",
+          )}
+        >
+          <h3 className="mb-2 mr-1 text-xs">{format(date, "EEE do")}</h3>
+          <DinnerSlot dinner={plannedDinner} />
+        </div>
+      </DialogTrigger>
+      <>
+        {changePlan || !plannedDinner ? (
+          <PlanDay
+            date={date}
+            closeDialog={() => onOpenChange(false)}
+            plannedDinner={plannedDinner}
+          />
+        ) : (
+          <PlannedDinner
+            dinner={plannedDinner}
+            date={date}
+            closeDialog={() => onOpenChange(false)}
+            setChangePlan={setChangePlan}
+          />
         )}
-      >
-        <h3 className="mb-2 mr-1 text-xs">{format(date, "EEE do")}</h3>
-        <DinnerSlot dinner={plannedDinner} />
-      </div>
-    </DialogTrigger>
+      </>
+    </Dialog>
   );
 };
 
@@ -28,17 +59,11 @@ type DinnerSlotProps = { dinner?: Dinner };
 
 export const DinnerSlot = ({ dinner }: DinnerSlotProps) => {
   if (!dinner) {
-    return <div className="h-12 rounded-md "></div>;
+    return <div className="h-12 rounded-md"></div>;
   }
 
-  return <DinnerSelected dinner={dinner} />;
-};
-
-type DinnerSelectedProps = { dinner: Dinner };
-
-const DinnerSelected = ({ dinner }: DinnerSelectedProps) => {
   return (
-    <div className={cn("flex h-12 flex-col-reverse rounded-md p-1")}>
+    <div className={cn("flex h-12 rounded-md p-1")}>
       <p className={cn("font-semibold")}>{dinner.name}</p>
     </div>
   );
