--- src/views/Plan/PlanDayDialog.tsx
@@ -1,96 +1,0 @@
-import { DialogDescription } from "@radix-ui/react-dialog";
-import { Button } from "../../components/ui/button";
-import {
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "../../components/ui/dialog";
-import { DialogDinners } from "./DialogDinners";
-import { api } from "../../utils/api";
-import { type Dinner } from "@prisma/client";
-import { usePostHog } from "posthog-js/react";
-import { format, isSameDay, startOfWeek } from "date-fns";
-
-type Props = {
-  date?: Date;
-  plannedDinner?: Dinner;
-  closeDialog: () => void;
-};
-
-export const PlanDayDialog = ({ date, plannedDinner, closeDialog }: Props) => {
-  const posthog = usePostHog();
-  const utils = api.useUtils();
-  const unplanDinnerMutation = api.plan.unplanDay.useMutation({
-    onMutate: (input) => {
-      void utils.plan.plannedDinners.cancel();
-      const prevPlannedDinners = utils.plan.plannedDinners.getData();
-      utils.plan.plannedDinners.setData(
-        { startOfWeek: startOfWeek(date ?? new Date(), { weekStartsOn: 1 }) },
-        (old) => {
-          return {
-            plans:
-              old?.plans.filter((plan) => !isSameDay(plan.date, input.date)) ??
-              [],
-          };
-        },
-      );
-      return { prevPlannedDinners };
-    },
-    onError: (_, __, context) => {
-      if (context?.prevPlannedDinners) {
-        utils.plan.plannedDinners.setData(
-          { startOfWeek: startOfWeek(date ?? new Date(), { weekStartsOn: 1 }) },
-          context.prevPlannedDinners,
-        );
-      }
-    },
-    onSettled: async () => {
-      await utils.plan.plannedDinners.invalidate();
-    },
-    onSuccess: (res) => {
-      posthog.capture("clear day", { day: format(res.deleted.date, "EEE do") });
-      closeDialog();
-    },
-  });
-
-  if (!date) {
-    return null;
-  }
-
-  const handleClear = () => {
-    unplanDinnerMutation.mutate({
-      date: date,
-      secret: localStorage.getItem("sulten-secret"),
-    });
-  };
-
-  return (
-    <DialogContent className="max-h-full">
-      <DialogHeader>
-        <DialogTitle>{format(date, "EEE do")}</DialogTitle>
-        <DialogDescription>
-          {plannedDinner ? plannedDinner.name : "Nothing planned yet"}
-        </DialogDescription>
-      </DialogHeader>
-      <div>
-        {date && (
-          <DialogDinners
-            date={date}
-            plannedDinner={plannedDinner}
-            closeDialog={closeDialog}
-          />
-        )}
-      </div>
-      <DialogFooter>
-        <Button
-          variant={"secondary"}
-          onClick={handleClear}
-          disabled={!plannedDinner}
-        >
-          Clear day
-        </Button>
-      </DialogFooter>
-    </DialogContent>
-  );
-};
