+++ src/views/Plan/PlanView.tsx
@@ -1,23 +1,21 @@
 import { UtensilsCrossed } from "lucide-react";
-import { Dialog } from "../../components/ui/dialog";
 import { api } from "../../utils/api";
-import { BottomNav } from "../BottomNav";
-import { Day } from "./Day";
-import { PlanDayDialog } from "./PlanDayDialog";
 import { useState } from "react";
 import { addDays, isSameDay, startOfDay, startOfWeek } from "date-fns";
+import { BottomNav } from "../BottomNav";
+import { Day } from "./Day";
 import { WeekSelect } from "../WeekSelect";
 
-export const WeekView = () => {
+export const PlanView = () => {
   const [selectedDay, setSelectedDay] = useState<Date>();
   const [dialogOpen, setDialogOpen] = useState(false);
-  const [weekOfSet, setWeekOfSet] = useState(0);
+  const [weekOffSet, setWeekOffSet] = useState(0);
   const trpc = api.useUtils();
 
   const startOfCurrentWeek = startOfWeek(new Date(), {
     weekStartsOn: 1,
   });
-  const startOfDisplayedWeek = addDays(startOfCurrentWeek, weekOfSet * 7);
+  const startOfDisplayedWeek = addDays(startOfCurrentWeek, weekOffSet * 7);
 
   const week: Date[] = [
     startOfDay(startOfDisplayedWeek),
@@ -60,41 +58,27 @@ export const WeekView = () => {
 
   return (
     <div className="grid h-screen">
-      <Dialog open={dialogOpen} onOpenChange={setDialogOpen}>
-        <div className="w-full space-y-4 p-4 ">
-          <WeekSelect
-            setWeekOfSet={setWeekOfSet}
-            startOfDisplayedWeek={startOfDisplayedWeek}
-          />
-          {week.map((day) => (
-            <Day
-              key={day.toString()}
-              date={day}
-              setSelectedDay={setSelectedDay}
-              plannedDinner={
-                plannedDinnersQuery.data?.plans.find((p) =>
-                  isSameDay(p.date, day),
-                )?.dinner
-              }
-            />
-          ))}
-          <WeekSelect
-            setWeekOfSet={setWeekOfSet}
-            startOfDisplayedWeek={startOfDisplayedWeek}
-          />
-        </div>
-        <PlanDayDialog
-          date={selectedDay}
-          plannedDinner={
-            selectedDay
-              ? plannedDinnersQuery.data?.plans.find((p) =>
-                  isSameDay(p.date, selectedDay),
-                )?.dinner
-              : undefined
-          }
-          closeDialog={() => setDialogOpen(false)}
+      <div className="w-full space-y-4 p-4 ">
+        <WeekSelect
+          setWeekOfSet={setWeekOffSet}
+          startOfDisplayedWeek={startOfDisplayedWeek}
         />
-      </Dialog>
+        {week.map((day) => (
+          <Day
+            key={day.toString()}
+            date={day}
+            plannedDinner={
+              plannedDinnersQuery.data?.plans.find((p) =>
+                isSameDay(p.date, day),
+              )?.dinner
+            }
+          />
+        ))}
+        <WeekSelect
+          setWeekOfSet={setWeekOffSet}
+          startOfDisplayedWeek={startOfDisplayedWeek}
+        />
+      </div>
 
       <BottomNav />
     </div>
